Move query logic into `WorkflowRepository`
