+++ packages/cli/src/constants.ts
@@ -85,8 +85,3 @@ export const LICENSE_QUOTAS = {
 export const UNLIMITED_LICENSE_QUOTA = -1;
 
 export const CREDENTIAL_BLANKING_VALUE = '__n8n_BLANK_VALUE_e5362baf-c777-4d57-a609-6eaf1f9e87f6';
-
-/**
- * Max items to return in a paginated list.
- */
-export const MAX_ITEMS = 50;
